+++ CHANGELOG.md
@@ -14,7 +14,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 --------------------
 * Bugs
   * Fixed F10.7 prelim and daily metadata to allow the fill value to keep the
-     same type as the data
+    same type as the data
 * Maintenance
   * Updated the GitHub action version numbers
   * Updated syntax for pysat instrument testing suite